6-size.c

+++ 0x00-hello_world/6-size.c
@@ -6,10 +6,10 @@
 */
 int main(void)
 {
-printf("Size of a char: %lu byte(s)\n", sizeof(char));
-prtintf("Size of an int: %lu byte(s)\n", sizeof(int));
-printf("Size of a long int: %lu byte(s)\n", sizeof(long int));
-printf("Size of a long long int: %lu byte(s)\n", sizeof(long long int));
-printf("Size of a float: %lu byte(s)\n", sizeof(float));
+printf("Size of a char: %u byte(s)\n", (unsigned long)sizeof(char));
+prtintf("Size of an int: %u byte(s)\n", (unsigned long)sizeof(int));
+printf("Size of a long int: %u byte(s)\n", (unsigned long)sizeof(long int));
+printf("Size of a long long int: %u byte(s)\n", (unsigned long)sizeof(long long int));
+printf("Size of a float: %u byte(s)\n", (unsigned long)sizeof(float));
 return (0);
 }
